+++ xarray_batcher/.gitignore
@@ -1,0 +1,5 @@
+__pycache__
+*.egg-info
+.flake8
+.pre-commit-config.yam
+.ipynb_checkpointsl
